Align `SchemaExtenderTest` with `graphql-js` v14.7
Tests no longer present in `graphql-js` has been moved to `SchemaExtenderLegacyTest`.
Respective commits that removed them from `extendSchema-test.js` are mentioned in comments.

Tests fail because:
- `repetable` in directives is lost when calling `extend()`
- `Schema::getAstNode()` returns `null` for an extended schema
- scalar types are added to schema automatically even if they are not used in SDL
